fix(mug): widen bucket math; enforce 0700 on load/list_users
- PooledExtractor::extract uses a u128 intermediate for i*dim bucketing so a
  large misconfigured dim can't overflow usize and panic on index
- EnrollStore::load and list_users call ensure_dir() so a pre-existing/loosened
  store dir is tightened back to 0700 before reading

Refs #57

+++ crates/mug/src/matcher.rs
@@ -128,8 +128,9 @@ impl EmbeddingExtractor for PooledExtractor {
         let mut counts = vec![0u64; self.dim];
         let n = bytes.len();
         for (i, &p) in bytes.iter().enumerate() {
-            // Contiguous segmentation of the flattened frame into `dim` buckets.
-            let bucket = (i * self.dim) / n;
+            // Contiguous segmentation of the flattened frame into `dim` buckets. Use a u128
+            // intermediate so a large `dim` (misconfigured) can't overflow `i * dim`.
+            let bucket = ((i as u128 * self.dim as u128) / n as u128) as usize;
             sums[bucket] += p as f64;
             counts[bucket] += 1;
         }
